added the requested changes
