+++ resources/lang/en/messages.php
@@ -1,0 +1,14 @@
+<?php
+
+return [
+    'success' => [
+        'create'        => 'Successfully saved',
+        'update'        => 'Successfully updated',
+        'delete'        => 'Successfully deleted',
+    ],
+    'error' => [
+        'create'        => 'Failed to save',
+        'update'        => 'Failed to update',
+        'delete'        => 'Failed to delete',
+    ],
+];
