+++ resources/views/admin/users.blade.php
@@ -66,6 +66,7 @@
                             @endif
                         </tbody>
                     </table>
+                    {{ $users->links('pagination::bootstrap-4') }}
                     </div>
                     </div>
                 </div>
